switched out os.path with pathlib; moved functions around
